requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmdstanpy==1.3.0
contourpy==1.3.2
copulae==0.8.0
-e git+ssh://git@github.com/RRemixx/cp-trajectory.git@ec28925333748f13b09a02eb7e9827b3ee36369e#egg=core
cycler==0.12.1
Cython==3.2.4
datasets==4.5.0
dill==0.4.0
epiweeks==2.4.0
exceptiongroup==1.3.1
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
holidays==0.88
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
hyperopt==0.2.7
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.15.0
networkx==3.4.2
numpy==2.2.6
online-conformal==1.0.2
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.0
plotly==6.5.2
propcache==0.4.1
prophet==1.2.1
py4j==0.10.9.9
pyarrow==22.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
salesforce-merlion==2.0.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stanio==0.5.1
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
